+++ models/delete.go
@@ -2,7 +2,7 @@ package models
 
 import "github.com/Lucasmartinsn/ApiProjeto-da-tropa/db"
 
-func Delete(id int) ( int, error){
+func Delete(id int64) ( int64, error){
 	conn, err := db.OpenConnection()
 	if err != nil{
 		return 0, err
